+++ RcppTskit/NEWS.md
@@ -35,6 +35,11 @@ and releases adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
   to append node rows from \code{R}, mirroring `tsk_node_table_add_row()`.
 - Added `rtsk_edge_table_add_row()` and `TableCollection$edge_table_add_row()`
   to append edge rows from \code{R}, mirroring `tsk_edge_table_add_row()`.
+- Added `rtsk_site_table_add_row()` and `TableCollection$site_table_add_row()`
+  to append site rows from \code{R}, mirroring `tsk_site_table_add_row()`.
+- Added `rtsk_mutation_table_add_row()` and
+  `TableCollection$mutation_table_add_row()` to append mutation rows from
+  \code{R}, mirroring `tsk_mutation_table_add_row()`.
 - TODO
 
 ### Changed